Add files via upload
Upload this file to the configuration files folder. The file must be named printer.cfg. This is the basic version to run and print on this printer. You can modify it as you like and I recommend adding more protections. Here you will find notes in Czech. I didn't want to translate it ....  sorry :)